Implement Proxmox Groups and Users
Since a new user is being created with administrator permissions, the root password setting on the end of the initialize_proxmox playbook can be removed.
root user will no longer be the main account to be used in development in Proxmox. This benefits security as well since root is predictable.