Mozilla SOPS KMS key rotation
Modified the .sops.yaml config file to utilize the newly created AWS KMS
keys for SOPS. Each .sops.* file key has been updated utilizing the
newly created bash script in the /hack/ dir. The script went through all
sops files and updated the keys to use the KMS keys instead of the local
age key. Additionally, modified the .gitattributes file to allow for
proper git diff functionality within the repo. This is still not working
as expected however since the .enc is required for proper diff, which
shows the something is misconfigured within the .gitconfig.