terraform .editorconfig formatting fix and license change
Modified the .editorconfig to follow two space formatting for .tf files.
This modified all tf file spacing structure to match. Additionally,
modified the LICNESE to include my name. Also modified the lint file to
include copypaste again since it is most likely best to keep that in and
create a shared hack/ folder for shared resources.